feat(kustomize): Update ml-workspace options
Updates the ml-workspace options in the drop-down to remove the root cpu and gpu ml-workspace images and replace them with the latest non-root cpu ml-workspace image, with rstudio, for next week's hackathon. If a different tag is more appropriate, let me know. This image will eventually be replaced with one or more that incorporate fixes [requested in the minimization PR](https://github.com/StatCan/kubeflow-containers/pull/34), but that is out of scope for the hackathon.