fix(ml-workspace-rstudio): fix disappearing tools
Updates the ml-workspace-rstudio image with a [hotfix](https://github.com/StatCan/kubeflow-desktop/pull/1) for the disappearing "Open Tools" dropdown. The issue was that whenever a Workspace Volume was mounted, the home directory was overwritten, wiping the configurations within. In this fix, the contents of the home directory are copied out and pasted back in.