feat(notebook): update remote-desktop
Includes the conda install fix, Mac support for clipboard sharing in Chrome (resolves https://github.com/StatCan/kubeflow-containers/issues/56 and resolves https://github.com/StatCan/kubeflow-containers/issues/57 if my understanding is correct that they come down to the same issue), and miscellaneous minor fixes. Does not include the landing page as that is still being finalized, but it should be ready for the next update.