Up the RAM and Workspace Volume notebook defaults
The nodes have a 1:4 CPU:RAM ratio, which we are not currently keeping with. (I'm guessing that'll lead to underutilized RAM?)

Also, I have heard reports of Jupyter kernels dying silently, which I am blaming the RAM for (though I do not have evidence.)

The workspace volume bump is partly because I think 10gb is a bit small (and it should be increased to at least 16gb), but also, we're exploring moving conda installs into the workspace volume, so we need space for conda.